Support custom user principal and home set paths
Currently, the user principal path and the home set path are both
hardcoded to "/", for both CalDAV and CardDAV. This poses a challenge if
one wishes to run a CardDAV and CalDAV server in the same server.

This commit adds a bit plumbing to offer a convenience function,
`ServeListHomeSets`, that can serve PROPFIND responses listing both
CalDAV and CardDAV home-set properties (calendar-home-set and
addressbook-home-set respectively).

To support this, the server must be aware of the current user's
principal URL. This is not a concern of either CalDAV or CardDAV, but
WebDAV in general. This commit allows an implementation to supply this
URL through the request context. If not set, the previous convention of
using "/" remains in place.

The individual servers will continue to work as before (including the
option of keeping everything at "/"). If one wishes to run CardDAV and
CalDAV in parallel, the new `webdav.ServeListHomeSets()` can be used to
serve requests for the current user's principal URL. The request routing
is the implementation's responsibility.

For added convenience, the `ServeListHomeSets()` function includes
support for discovering the home sets via their `/.well-known` URLs.

The home set path has to be provided by the storage backend. It is
customary, though not required, that the home set path includes the
user's pricipal path. The storage backends already have access to the
request context, which allows computing the home set path based on the
pricipal path.