carddav: fix PROPFIND for all path layouts
Refactor the PROPFIND handling to properly support all possible path
layouts (from "everything at /" to "everything at different path"). This
includes adding a new method the Backend interface, as it is the
authority that determines the path layout.

This fixes the recent regression of the broken "everything at /" layout.
It also already provides the needed bits and pieces to support multiple
address books.