caldav: support more features for recurring events
The protocol allows clients to ask the server to e.g. expand recurring
events to a set of simple events in a given time-frame, so that the
client does not have to calculate the recurrences.

This commit adds the foundation for support for expand [1],
limit-recurrence-set [2], and limit-freebusy-set [3] functionality.
However, the actual transformation of events returned to clients will
have to be implemented in the backend, which now has the required
information to do so.

[1]: https://datatracker.ietf.org/doc/html/rfc4791#section-9.6.5
[2]: https://datatracker.ietf.org/doc/html/rfc4791#section-9.6.6
[3]: https://datatracker.ietf.org/doc/html/rfc4791#section-9.6.7