fix: handle "allprop" with XML content type
The previous code failed to properly handle PROPFIND with an empty
request body (RFC 4918 section 9.1) for XML content type requests.

Refactor the logic to always consider PROPFIND requests with empty
request bodies as "allprop" requests, regardless of content type.

There is no need to worry about the request's content type if the
request body is empty anyway, so only parse if we know it is not.